Improve readibility for SvgParser commandMap
